EVEREST-633 install monitoring stack with install command
By moving to a multi-namespace everest-operator installation we
configured the operator group of the percona-everest namespace to have N
target namespaces. The VM Operator doesn't support multi-namespaces so
we can't deploy it to the same namespace as the everest-operator.
Therefore we install it onto a separate namespace for managing
everything monitoring related.