EVEREST-633 Install monitoring stack onto a separate namespace (#281)
* EVEREST-633 install monitoring stack with install command

By moving to a multi-namespace everest-operator installation we
configured the operator group of the percona-everest namespace to have N
target namespaces. The VM Operator doesn't support multi-namespaces so
we can't deploy it to the same namespace as the everest-operator.
Therefore we install it onto a separate namespace for managing
everything monitoring related.

* EVEREST-633 remove monitoring command

* EVEREST-633 Update everest-operator go mod

* EVEREST-633 Update percona-everest-backend go mod

* EVEREST-633 forbid installs in the monitoring namespace

* EVEREST-633 update namespaces env var name

* EVEREST-633 Update everest-operator go mod