EVEREST-495 Add multi namespaces (#283)
* EVEREST-633 Add multi-namespace support (#271)

* EVEREST-633 force everest install to percona-everest namespace

* EVEREST-633 remove deprecated name flag

* EVEREST-633 add multi-namespace support

* EVEREST-633 remove operator channel flags

* EVEREST-633 update backend and operator go mods

* EVEREST-633 shut up linter, we know this code is dead

* EVEREST-633 Install monitoring stack onto a separate namespace (#281)

* EVEREST-633 install monitoring stack with install command

By moving to a multi-namespace everest-operator installation we
configured the operator group of the percona-everest namespace to have N
target namespaces. The VM Operator doesn't support multi-namespaces so
we can't deploy it to the same namespace as the everest-operator.
Therefore we install it onto a separate namespace for managing
everything monitoring related.

* EVEREST-633 remove monitoring command

* EVEREST-633 Update everest-operator go mod

* EVEREST-633 Update percona-everest-backend go mod

* EVEREST-633 forbid installs in the monitoring namespace

* EVEREST-633 update namespaces env var name

* EVEREST-633 Update everest-operator go mod

* EVEREST-633 Use PGO stable channel (#282)

* EVEREST-633 install monitoring stack with install command

By moving to a multi-namespace everest-operator installation we
configured the operator group of the percona-everest namespace to have N
target namespaces. The VM Operator doesn't support multi-namespaces so
we can't deploy it to the same namespace as the everest-operator.
Therefore we install it onto a separate namespace for managing
everything monitoring related.

* EVEREST-633 remove monitoring command

* EVEREST-633 Update everest-operator go mod

* EVEREST-633 Update percona-everest-backend go mod

* EVEREST-633 forbid installs in the monitoring namespace

* EVEREST-633 use PGO stable channel

* EVEREST-495 Update everest-operator go mod

* EVEREST-495 Fix missing monitoring namespace env var