EVEREST-838 Workaround waiting for DBs to be deleted (#311)
* EVEREST-838 workaround waiting for DBs to be deleted

When deleting a DBC CR, the everest operator doesn't wait for the DB
operator's CRs to be deleted. Thus, as soon as we delete the DBC CRs,
these cease to exist in the cluster and the polling below will return
immediately. If we don't wait for the DB operators to process the
deletion of the CRs, we may end up deleting the namespaces before the DB
operators have a chance to delete the resources they manage, leaving the
namespaces in an endless Terminating state waiting for finalizers to be
removed.
The everest operator should have a Deleting status that waits for the DB
operators to delete their DB CRs before removing the corresponting DBC
CR. Until this is implemented, we work around this by sleeping for two
minutes to give the DB operators a chance to delete the resources they
manage before we delete the namespaces.

* EVEREST-838 Update everest-operator go mod

* EVEREST-838 increase tests timeout to 10 min

With the recent changes that introduced the install of the monitoring
stack by default, the install command now takes longer so we shall
increase the test timeout to avoid intermittent CI failures.

* EVEREST-838 fix cli-tests with new uninstall command